LayoutTests/platform/wpe/http/wpt/webxr/xrDevice_requestSession_previously_enabled_feature_not_requested.https-expected.txt

CONSOLE MESSAGE: Error: assert_equals: expected (object) null but got (undefined) undefined

Harness Error (FAIL), message = Error: assert_equals: expected (object) null but got (undefined) undefined

TIMEOUT requestSession should not grant previously enabled features if not requested - webgl Test timed out
NOTRUN requestSession should not grant previously enabled features if not requested - webgl2
